requirements.txt
accelerate==1.10.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.19.1
fsspec==2025.7.0
gitdb==4.0.12
gitpython==3.1.45
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
markupsafe==3.0.2
mpmath==1.3.0
narwhals==2.2.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
peft==0.17.1
pillow==11.3.0
pip==25.2
portalocker==3.2.0
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pymupdf==1.26.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.15.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rpds-py==0.27.1
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.49.0
sympy==1.14.0
tenacity==9.1.2
timm==1.0.19
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.55.4
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
